+++ app-mall-api/src/main/java/app/mall/api/beans/LitemallBrandInputBean.java
@@ -54,16 +54,16 @@
         }
 
 
-        private Byte _sortOrder;
+        private Integer _sortOrder;
 
     
         @PropMeta(propId=5)
     
-        public Byte getSortOrder(){
+        public Integer getSortOrder(){
             return _sortOrder;
         }
 
-        public void setSortOrder(Byte value){
+        public void setSortOrder(Integer value){
             this._sortOrder = value;
         }
